+++ src/main/java/org/torproject/collector/bridgedescs/BridgeDescriptorParser.java
@@ -27,7 +27,7 @@ public class BridgeDescriptorParser {
 
   /** Parses the first line of the given descriptor data to determine the
    * descriptor type and passes it to the sanitized bridges writer. */
-  public void parse(byte[] allData, String dateTime)
+  public void parse(byte[] allData, String dateTime, String authorityFingerprint)
       throws ConfigurationException {
     try {
       BufferedReader br = new BufferedReader(new StringReader(
@@ -45,7 +45,8 @@ public class BridgeDescriptorParser {
         }
       } else {
         if (this.sbw != null) {
-          this.sbw.sanitizeAndStoreNetworkStatus(allData, dateTime);
+          this.sbw.sanitizeAndStoreNetworkStatus(allData, dateTime,
+              authorityFingerprint);
         }
       }
     } catch (IOException e) {